reorganized services, deps and mocks
