+++ server/app.py
@@ -3,7 +3,7 @@
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
-from server.deps.mocks.deps_overrides import DepsOverrides
+from server.deps_overrides import DepsOverrides
 from server.exception_handlers import add_exception_handlers
 from server.routes.admin import router as AdminRouter
 from server.routes.public import router as PublicRouter
